move branch functions to modules to avoid dependencies cycles
